+++ src/Sentry.Unity/SentryUnityOptions.cs
@@ -155,6 +155,16 @@ namespace Sentry.Unity
         public bool AndroidNativeSupportEnabled { get; set; } = true;
 
         /// <summary>
+        /// Whether the SDK should add the NDK integration for Android
+        /// </summary>
+        public bool NdkIntegrationEnabled { get; set; } = true;
+
+        /// <summary>
+        /// Whether the SDK should sync the scope to the NDK layer for Android
+        /// </summary>
+        public bool NdkScopeSyncEnabled { get; set; } = true;
+
+        /// <summary>
         /// Whether the SDK should add native support for Windows
         /// </summary>
         public bool WindowsNativeSupportEnabled { get; set; } = true;